requirements.txt
discord
asyncio
pillow
requests
regex
yt_dlp
nest-asyncio
pynacl
mcstats
pandas